save and load
--- project ---
main.rb -- made some small adjustments to the inventory and notes
setup, with the intention of making things more useful. Saving should
now be completed. Loading is now mostly complete, with only the
content of inventory not being loaded yet.

--- notes ---
Eventually you should be able to have the character information fully
editable without any need of using the attribute readers. After that,
the project should have reached a "completed" state, but will not be
the end of this project. This will eventually be a Rails project, so
will have to look into how to set up that.